requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1
django-celery-beat==2.5.0
APScheduler==3.10.4
gunicorn==21.2.0
dj-database-url==2.1.0
setuptools==80.9.0
